+++ lib/map.dart
@@ -27,3 +27,4 @@ export 'src/map/neshan_map.dart';
 export 'src/map/config/neshan_map_config.dart';
 export 'src/map/controller/neshan_map_controller.dart';
 export 'src/map/models/neshan_marker.dart';
+export 'src/utils/neshan_common.dart';